The most correct results
I have deleted some functions that is not necessary for the harmonic
potential case. Also, I have taken out the scale constant because it
has been the reason why the program produces unstable results. I have
added the constant omega as the new and correct scaling constant. With
that the program now produces the correct results for stationary
initial conditions and superpositions of states. NOTE: THE PROGRAM IS
USING ONLY ONE MASS AND THEREFORE I HAVE DEFINED ONLY ONE OMEGA. IF
NEEDED FOR SIMULATING A SECOND MASS, ONE NEEDS TO DEFINE A SECOND
VARIABLE FOR MASS AS WELL AS A SECOND OMEGA.